Revert "타인의 꿈 페이지 완성"
