+++ .gitignore
@@ -59,5 +59,3 @@ buck-out/
 # Ruby / CocoaPods
 /ios/Pods/
 /vendor/bundle/
-
-.env
