+++ App.tsx
@@ -10,9 +10,6 @@ import AdvicePage from './src/pages/Dream/AdvicePage'
 import ShareDream from './src/pages/Dream/ShareDream'
 import TravelPage from './src/pages/Dream/TravelPage'
 import WritingPage from './src/pages/Dream/WritingPage'
-import MyProfilePage from './src/pages/My/MyProfilePage'
-import LoginPage from './src/pages/Login/LoginPage'
-import RegisterPage from './src/pages/Login/RegisterPage'
 import ResultPage from './src/pages/Dream/ResultPage'
 
 const Stack = createNativeStackNavigator();
@@ -29,9 +26,6 @@ export default function App() {
         <Stack.Screen name="ShareDream" component={ShareDream} />
         <Stack.Screen name="TravelPage" component={TravelPage} />
         <Stack.Screen name="WritingPage" component={WritingPage} />
-        <Stack.Screen name="MyProfilePage" component={MyProfilePage} />
-        <Stack.Screen name="LoginPage" component={LoginPage} />
-        <Stack.Screen name="RegisterPage" component={RegisterPage} />
         <Stack.Screen name="ResultPage" component={ResultPage} />
       </Stack.Navigator>
     </NavigationContainer>
