--- android/app/src/main/res/xml/network_security_config.xml
@@ -1,5 +1,0 @@
-<network-security-config>
-    <domain-config cleartextTrafficPermitted="true">
-        <domain includeSubdomains="true">10.0.2.2</domain>
-    </domain-config>
-</network-security-config>
