+++ android/gradle.properties
