+++ package.json
@@ -17,7 +17,6 @@
     "@react-navigation/stack": "^6.2.1",
     "react": "17.0.2",
     "react-native": "^0.68.2",
-    "react-native-config": "^1.4.6",
     "react-native-gesture-handler": "^2.4.2",
     "react-native-gradle-plugin": "^0.0.7",
     "react-native-paper": "^4.12.1",
