+++ src/pages/Dream/AdvicePage.tsx
@@ -73,11 +73,9 @@ const styles = StyleSheet.create({
     marginTop: 30,
     backgroundColor: '#DDCEFF',
     borderRadius: 5,
-    opacity: 0.7,
   },
   travelText: {
     margin: 5,
-    alignSelf: 'center',
   },
   BottomFlowerView: {
     flexDirection: 'row',
