+++ src/pages/Dream/TravelPage.tsx
@@ -76,10 +76,8 @@ const styles = StyleSheet.create({
     marginTop: 30,
     backgroundColor: '#DDCEFF',
     borderRadius: 5,
-    opacity: 0.7,
   },
   travelText: {
     margin: 5,
-    alignSelf: 'center',
   },
 })
