+++ src/pages/My/Mypage.tsx
@@ -43,23 +43,6 @@ export default function MyPage({navigation}) {
       <TouchableHighlight
         style={styles.button}
         onPress={() => {
-          navigation.navigate('MyProfilePage', {screen: 'MyProfilePage'})
-        }}>
-        <Text>profile page</Text>
-      </TouchableHighlight>
-      <TouchableHighlight
-        style={styles.button}
-        onPress={() => {
-          navigation.navigate('LoginPage', {screen: 'LoginPage'})
-        }}>
-        <Text>login page</Text>
-      </TouchableHighlight>
-      <TouchableHighlight
-        style={styles.button}
-        onPress={() => {
-          navigation.navigate('RegisterPage', {screen: 'RegisterPage'})
-        }}>
-        <Text>register page</Text>
           navigation.navigate('ResultPage', {screen: 'ResultPage'})
         }}>
         <Text>result page</Text>
